feat:  Implementa módulo completo de viagens (CRUD + status)
Adiciona entidade Viagem, value object Localizacao e enum StatusViagemEnum. Implementa comandos, queries, handlers e repositório para criação, alteração de status e listagem de viagens. Inclui controller REST, validação com FluentValidation, mapeamento EF Core e migration. Adiciona testes automatizados para todos os cenários. Ajusta DI e tratamento global de exceções.

+++ tests/Fleet.Application.Tests/Condutores/CondutorTestFactory.cs
@@ -10,8 +10,8 @@ namespace Fleet.Application.Tests.Condutores;
 /// </summary>
 internal static class CondutorTestFactory
 {
-    private static int _cpfCounter = 10000000000;
-    private static int _cnhCounter = 10000000000;
+    private static long _cpfCounter = 10000000000L;
+    private static long _cnhCounter = 10000000000L;
 
     /// <summary>
     /// Cria um condutor com valores padrão, permitindo sobrescrita de campos específicos.
